Moving the kafka test code to test folder to avoid conflicts with the kafka under li layer
