+++ build.gradle
@@ -135,12 +135,10 @@ project(':datastream-testcommon') {
     compile project(':datastream-server-api')
     compile project(':datastream-common')
     compile project(':datastream-utils')
-    compile project(':datastream-kafka')
     compile project(':datastream-server')
     compile "commons-cli:commons-cli:$commonsCliVersion"
     compile "org.apache.avro:avro:$avroVersion"
     compile "org.apache.zookeeper:zookeeper:$zookeeperVersion"
-    compile "org.apache.kafka:kafka_2.10:$kafkaVersion"
     compile "com.linkedin.pegasus:restli-server:$pegasusVersion"
   }
 }
@@ -177,6 +175,8 @@ project(':datastream-server') {
     compile project(':datastream-common')
     compile project(':datastream-utils')
 
+    testCompile project(':datastream-kafka')
+    testCompile "org.apache.kafka:kafka_2.10:$kafkaVersion"
     testCompile project(':datastream-client')
     testCompile project(':datastream-file-connector')
     testCompile project(':datastream-testcommon')
