+++ datastream-server/src/test/java/com/linkedin/datastream/server/dms/TestBootstrapActionResources.java
@@ -3,7 +3,7 @@ package com.linkedin.datastream.server.dms;
 import com.linkedin.datastream.common.Datastream;
 import com.linkedin.datastream.server.TestDatastreamServer;
 import com.linkedin.datastream.connectors.DummyBootstrapConnector;
-import com.linkedin.datastream.testutil.EmbeddedDatastreamCluster;
+import com.linkedin.datastream.server.EmbeddedDatastreamCluster;
 import com.linkedin.restli.server.RestLiServiceException;
 import org.testng.Assert;
 import org.testng.annotations.AfterMethod;
